+++ packages/server/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @hot-updater/server
 
+## 0.22.0
+
+### Minor Changes
+
+- 32ad614: feat(server): integrate endpoint `/bundles/*` => `/api/bundles/*`
+
+### Patch Changes
+
+- @hot-updater/core@0.22.0
+- @hot-updater/js@0.22.0
+- @hot-updater/plugin-core@0.22.0
+
 ## 0.21.15
 
 ### Patch Changes
@@ -126,7 +138,7 @@
   - @hot-updater/plugin-core@0.21.1
   - @hot-updater/core@0.21.1
 
-## 1.0.0
+## 0.22.0
 
 ### Minor Changes
 
@@ -137,5 +149,5 @@
 - Updated dependencies [610b2dd]
 - Updated dependencies [afb084b]
 - Updated dependencies [036f8f0]
-  - @hot-updater/plugin-core@1.0.0
-  - @hot-updater/core@1.0.0
+  - @hot-updater/plugin-core@0.22.0
+  - @hot-updater/core@0.22.0