+++ packages/test-utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @hot-updater/test-utils
 
+## 0.22.0
+
+### Patch Changes
+
+- @hot-updater/core@0.22.0
+
 ## 0.21.15
 
 ### Patch Changes
@@ -96,7 +102,7 @@
 
 - @hot-updater/core@0.21.1
 
-## 1.0.0
+## 0.22.0
 
 ### Minor Changes
 
@@ -106,4 +112,4 @@
 
 - Updated dependencies [afb084b]
 - Updated dependencies [036f8f0]
-  - @hot-updater/core@1.0.0
+  - @hot-updater/core@0.22.0