+++ plugins/expo/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @hot-updater/expo
 
+## 0.22.0
+
+### Patch Changes
+
+- @hot-updater/cli-tools@0.22.0
+- @hot-updater/bare@0.22.0
+- @hot-updater/plugin-core@0.22.0
+
 ## 0.21.15
 
 ### Patch Changes
@@ -122,7 +130,7 @@
   - @hot-updater/plugin-core@0.21.1
   - @hot-updater/bare@0.21.1
 
-## 1.0.0
+## 0.22.0
 
 ### Minor Changes
 
@@ -133,8 +141,8 @@
 - Updated dependencies [610b2dd]
 - Updated dependencies [afb084b]
 - Updated dependencies [036f8f0]
-  - @hot-updater/plugin-core@1.0.0
-  - @hot-updater/bare@1.0.0
+  - @hot-updater/plugin-core@0.22.0
+  - @hot-updater/bare@0.22.0
 
 ## 0.20.15