+++ plugins/js/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @hot-updater/js
 
+## 0.22.0
+
+### Patch Changes
+
+- @hot-updater/core@0.22.0
+
 ## 0.21.15
 
 ### Patch Changes
@@ -92,7 +98,7 @@
 
 - @hot-updater/core@0.21.1
 
-## 1.0.0
+## 0.22.0
 
 ### Minor Changes
 
@@ -103,7 +109,7 @@
 
 - Updated dependencies [afb084b]
 - Updated dependencies [036f8f0]
-  - @hot-updater/core@1.0.0
+  - @hot-updater/core@0.22.0
 
 ## 0.20.15