+++ plugins/supabase/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @hot-updater/supabase
 
+## 0.22.0
+
+### Patch Changes
+
+- @hot-updater/cli-tools@0.22.0
+- @hot-updater/core@0.22.0
+- @hot-updater/plugin-core@0.22.0
+
 ## 0.21.15
 
 ### Patch Changes
@@ -124,7 +132,7 @@
   - @hot-updater/plugin-core@0.21.1
   - @hot-updater/core@0.21.1
 
-## 1.0.0
+## 0.22.0
 
 ### Minor Changes
 
@@ -137,8 +145,8 @@
 - Updated dependencies [610b2dd]
 - Updated dependencies [afb084b]
 - Updated dependencies [036f8f0]
-  - @hot-updater/plugin-core@1.0.0
-  - @hot-updater/core@1.0.0
+  - @hot-updater/plugin-core@0.22.0
+  - @hot-updater/core@0.22.0
 
 ## 0.20.15